rename file containing all time literals
